@@ -1,6 +1,6 @@
 module sigs.k8s.io/release-utils
 
-go 1.21
+go 1.23
 
 require (
 	github.com/blang/semver/v4 v4.0.0
